moved + button, refactored map screen
